tgService/pyproject.toml
[tool.poetry]
name = "tgservice"
version = "0.1.0"
description = ""
authors = ["German Legkovoy <147488090+legkovoyg@users.noreply.github.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.12"
aiogram = "^3.17.0"
python-dotenv = "^1.0.1"
requests = "^2.32.3"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
